backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
et_xmlfile==2.0.0
inflection==0.5.1
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
PyJWT==2.12.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
groq
striprtf
python-docx
pydantic
python-dotenv
